Clean RDMA folder and remove unused pingpongs.
The original rdma-core pingpongs were removed since they would not be used because of their synchronous design. They might still be useful for comparison and as a guideline to write new programs; however, the original code in `rdma-core` repository can be used for that purpose. The only difference was the documentation added in the code, which can always be found by looking in the git history of the file.
