--- rdma/pingpongs/src/pingpong.c
@@ -1,101 +1,0 @@
-/*
- * Copyright (c) 2006 Cisco Systems.  All rights reserved.
- *
- * This software is available to you under a choice of one of two
- * licenses.  You may choose to be licensed under the terms of the GNU
- * General Public License (GPL) Version 2, available from the file
- * COPYING in the main directory of this source tree, or the
- * OpenIB.org BSD license below:
- *
- *     Redistribution and use in source and binary forms, with or
- *     without modification, are permitted provided that the following
- *     conditions are met:
- *
- *      - Redistributions of source code must retain the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer.
- *
- *      - Redistributions in binary form must reproduce the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer in the documentation and/or other materials
- *        provided with the distribution.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#include "pingpong.h"
-#include <endian.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-inline enum ibv_mtu pp_mtu_to_enum (int mtu)
-{
-    switch (mtu)
-    {
-    case 256:
-        return IBV_MTU_256;
-    case 512:
-        return IBV_MTU_512;
-    case 1024:
-        return IBV_MTU_1024;
-    case 2048:
-        return IBV_MTU_2048;
-    case 4096:
-        return IBV_MTU_4096;
-    default:
-        return 0;
-    }
-}
-
-/**
- * Convert a wire gid string to a gid structure.
- *
- * @param wgid the wire gid string
- * @param gid the gid structure
- */
-void wire_gid_to_gid (const char *wgid, union ibv_gid *gid)
-{
-    char tmp[9];
-    __be32 v32;
-    int i;
-    uint32_t tmp_gid[4];
-
-    for (tmp[8] = 0, i = 0; i < 4; ++i)
-    {
-        memcpy (tmp, wgid + i * 8, 8);
-        sscanf (tmp, "%x", &v32);
-        tmp_gid[i] = be32toh (v32);
-    }
-    memcpy (gid, tmp_gid, sizeof (*gid));
-}
-
-/**
- * Convert a gid structure to a gid string to be transmitted.
- *
- * @param gid the gid structure
- * @param wgid the wire gid string
- */
-void gid_to_wire_gid (const union ibv_gid *gid, char wgid[])
-{
-    uint32_t tmp_gid[4];
-    int i;
-
-    memcpy (tmp_gid, gid, sizeof (tmp_gid));
-    for (i = 0; i < 4; ++i)
-        sprintf (&wgid[i * 8], "%08x", htobe32 (tmp_gid[i]));
-}
-
-long long get_nanos (void)
-{
-    struct timespec ts;
-    timespec_get (&ts, TIME_UTC);
-    return (long long) ts.tv_sec * 1000000000L + ts.tv_nsec;
-}
