+++ rdma/ud_pingpong.c
@@ -8,9 +8,9 @@
 #include <time.h>
 #include <unistd.h>
 
-#include "../../common/common.h"
-#include "../../common/net.h"
-#include "../../common/persistence.h"
+#include "../common/common.h"
+#include "../common/net.h"
+#include "../common/persistence.h"
 //#include "ccan/minmax.h"
 #include "src/ib_net.h"
 #include "src/pingpong.h"
